fix-copy-literal-bug (#1)
Closes: #265
